app/requirements.txt
Flask==2.3.3
psycopg2-binary==2.9.7
boto3==1.29.7
botocore==1.32.7
gunicorn==21.2.0
python-json-logger==2.0.7